Refactored classes, created placement_validation module
