+++ lib/game.rb
@@ -32,7 +32,7 @@ class Game
 
   def setup
     generate_boards
-    @cpu.setup
+    @cpu.alt_setup
     ship_info
     puts @player.board.render(true)
     @player.place_ships
@@ -43,16 +43,30 @@ class Game
     width = 4
     height = 4
     loop do
-      print "Select board width (minimum 4, maximum 26):\n>  "
-      width = gets.chomp.to_i
-      print "Select board height (minimum 4, maximum 26):\n>  "
-      height = gets.chomp.to_i
-      if (width < 4 || height < 4) || (width > 26 || height > 26)
-        puts "Sorry, those dimensions are invalid. Please enter a valid board size."
-      else
+      print "Would you like to choose the board size? (y/n)\n>  "
+      select = gets.chomp.downcase
+      if select == 'y' || select == 'yes'
+        loop do
+          print "Select board width (minimum 4, maximum 26):\n>  "
+          width = gets.chomp.to_i
+          print "Select board height (minimum 4, maximum 26):\n>  "
+          height = gets.chomp.to_i
+          if (width < 4 || height < 4) || (width > 26 || height > 26)
+            puts "Sorry, those dimensions are invalid. Please enter a valid board size."
+          else
+            break
+          end
+        end
         break
+      elsif select == 'n' || select == 'no'
+        puts "In that case, we'll play on a 4 x 4 board"
+        break
+      else
+        puts "I'm sorry that doesn't answer my question."
       end
-    end
+
+  end
+
     @player.customize_board(width, height)
     @cpu.customize_board(width, height)
   end
@@ -78,9 +92,9 @@ class Game
 
   def display_boards
     #variable '=' based on board size
-    puts "#{'=' * 14}COMPUTER BOARD==============="
+    puts "#{'=' * (@cpu.board.cells.values.last.coordinate[1..-1].to_i + 6)}COMPUTER BOARD#{'=' * (@cpu.board.cells.values.last.coordinate[1..-1].to_i + 6)}"
     puts @cpu.board.render
-    puts "================PLAYER BOARD================"
+    puts "#{'=' * (@cpu.board.cells.values.last.coordinate[1..-1].to_i + 7)}PLAYER BOARD#{'=' * (@cpu.board.cells.values.last.coordinate[1..-1].to_i + 7)}"
     puts @player.board.render(true)
     print "Enter the coordinate for your shot:\n>  "
     shots_fired
@@ -88,7 +102,6 @@ class Game
 
   def turn
     # Tracking turn count?
-    puts "We're all set to play!\n\n\n"
     loop do
       if end_game?
         end_game
