+++ runner.rb
@@ -1,9 +1,10 @@
+require './lib/ship'
+require './lib/cell'
+require './lib/placement_validation'
+require './lib/board'
 require './lib/computer'
 require './lib/player'
 require './lib/game'
-require './lib/board'
-require './lib/cell'
-require './lib/ship'
 
 game = Game.new
 game.main_menu
